Renamed class Optimization_Subscribers into Optimization\ServiceProvider Also created an alias and use `rocket_get_constant()`.
