+++ inc/deprecated/3.6.php
@@ -1,0 +1,9 @@
+<?php
+// phpcs:ignoreFile
+
+defined( 'ABSPATH' ) || exit;
+
+/**
+ * Class aliases.
+ */
+class_alias( '\WP_Rocket\Optimization\ServiceProvider', '\WP_Rocket\ServiceProvider\Optimization_Subscribers' );
