+++ inc/main.php
@@ -53,6 +53,7 @@ function rocket_init() {
 	require WP_ROCKET_DEPRECATED_PATH . '3.3.php';
 	require WP_ROCKET_DEPRECATED_PATH . '3.4.php';
 	require WP_ROCKET_DEPRECATED_PATH . '3.5.php';
+	require WP_ROCKET_DEPRECATED_PATH . '3.6.php';
 	require WP_ROCKET_3RD_PARTY_PATH . '3rd-party.php';
 	require WP_ROCKET_COMMON_PATH . 'admin-bar.php';
 	require WP_ROCKET_COMMON_PATH . 'emoji.php';
